ParfaitHR/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-filter==2.4.0
gunicorn==20.0.4
Pillow==8.0.1
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
